Converted SDK into multi-module project Moved older project under 'outbound-sdk' module Created new module for Elasticsearch Refactored ElasticsearchOutboundRecord and ElasticsearchOutboundMetadata
