+++ buildSrc/src/main/kotlin/com/aerospike/connect/VulnerabilityScanningExtensions.kt
@@ -1,0 +1,28 @@
+package com.aerospike.connect
+
+import io.snyk.gradle.plugin.SnykExtension
+import org.gradle.api.Project
+import org.gradle.api.tasks.Exec
+import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.create
+import org.gradle.kotlin.dsl.provideDelegate
+
+fun Project.setupVulnerabilityScanning() {
+    val snykTokens: String by project
+    val snykToken = snykTokens.split(",").map { it.trim() }.random()
+
+    tasks.create<Exec>("setup-snyk") {
+        commandLine("${project.rootDir}/snyk", "auth", snykToken)
+    }
+    tasks.getByName("snyk-check-binary").finalizedBy("setup-snyk")
+
+    /**
+     * Vulnerability scanning with Snyk.
+     */
+    configure<SnykExtension> {
+        setApi(snykToken)
+        setSeverity("high")
+        setAutoDownload(true)
+        setArguments("--sub-project=" + project.name)
+    }
+}
