+++ outbound-sdk/gradle.properties
