+++ outbound-sdk/src/main/java/com/aerospike/connect/outbound/AerospikeOperation.java
