+++ outbound-sdk/src/main/java/com/aerospike/connect/outbound/ParticleType.java
