+++ outbound-sdk/src/main/java/com/aerospike/connect/outbound/format/MediaType.java
