+++ outbound-sdk/src/main/java/com/aerospike/connect/outbound/format/OutboundRecord.java
