+++ outbound-sdk/src/main/java/com/aerospike/connect/outbound/format/SkipOutboundRecord.java
