+++ outbound-sdk/src/main/java/com/aerospike/connect/outbound/routing/RouterConfig.java
