+++ outbound-sdk/src/main/java/com/aerospike/connect/outbound/transformer/SkipChangeNotificationRecord.java
